@@ -1,8 +1,10 @@
 """
-Base classes for parallel model fitting. See the specific functions and classes in `tm_gensim`, `tm_lda` and
-`tm_sklearn` for parallel processing with popular topic modeling packages.
+Base classes for parallel model fitting and evaluation. See the specific functions and classes in
+:mod:`~tmtoolkit.topicmod.tm_gensim`, :mod:`~tmtoolkit.topicmod.tm_lda` and :mod:`~tmtoolkit.topicmod.tm_sklearn` for
+parallel processing with popular topic modeling packages.
 
-Markus Konrad <markus.konrad@wzb.eu>
+.. note:: The classes and functions in this module are only important if you want to implement your own parallel
+          model computation and evaluation.
 """
 
 
@@ -22,8 +24,28 @@ logger = logging.getLogger('tmtoolkit')
 #%% General parallel model computation
 
 
-class MultiprocModelsRunner(object):
+class MultiprocModelsRunner:
+    """
+    Runner class for distributing and managing worker processes for parallel model computation.
+    """
+
     def __init__(self, worker_class, data, varying_parameters=None, constant_parameters=None, n_max_processes=None):
+        """
+        Initiate runner class with a model computation worker class `worker_class` (which should be derived from
+        :class:`~tmtoolkit.topicmod.parallel.MultiprocModelsWorkerABC`). This class represents the worker
+        processes and each will be instantiated with `data` and work on it with a different parameter set that can be
+        passed via `varying_parameters`.
+
+        :param worker_class: model computation worker class derived from
+                             :class:`~tmtoolkit.topicmod.parallel.MultiprocModelsWorkerABC`
+        :param data: the data that the workers use for computations; 2D (sparse) array/matrix or a dict with such
+                     matrices; the latter allows to run all computations on different datasets at once
+        :param varying_parameters: list of dicts with parameters; each parameter set will be used in a separate
+                                   computation
+        :param constant_parameters: dict with parameters that are the same for all parallel computations
+        :param n_max_processes: maximum number of worker processes to spawn
+        """
+
         self.tasks_queues = None
         self.results_queue = None
         self.workers = None
@@ -46,6 +68,7 @@ class MultiprocModelsRunner(object):
         else:
             self.data = {None: self._prepare_data(data)}
 
+        # number of workers: at least as much as needed for the varying params and datasets but at max. n_max_processes
         self.n_workers = min(max(1, n_varying_params) * len(self.data), n_max_processes)
 
         logger.info('init with %d workers' % self.n_workers)
@@ -57,6 +80,7 @@ class MultiprocModelsRunner(object):
         self.shutdown_workers()
 
     def shutdown_workers(self):
+        """Send shutdown signal to all worker processes to stop them."""
         if not self.workers:
             return
 
@@ -73,10 +97,22 @@ class MultiprocModelsRunner(object):
         self.n_workers = 0
 
     def run(self):
+        """
+        Set up worker processes and run parallel computations. Blocks until all processes are done, then stops all
+        workers and returns the results.
+
+        :return: if passed data is 2D array, returns a list with tuples (parameter set, results); if passed data is
+                 a dict of 2D arrays, returns dict with same keys as data and the respective results for each dataset
+        """
+
+        # set up worker processes
         self._setup_workers(self.worker_class)
 
+        # merge varying and constant parameters to get parameter set for each worker
         params = _merge_params(self.varying_parameters, self.constant_parameters)
         n_params = len(params)
+
+        # create the worker tasks: cartesian product of parameter sets and datasets (here named "docs")
         docs = list(self.data.keys())
         n_docs = len(docs)
         if n_params == 0:
@@ -84,9 +120,11 @@ class MultiprocModelsRunner(object):
         else:
             tasks = list(itertools.product(docs, params))
         n_tasks = len(tasks)
+
         logger.info('multiproc models: starting with %d parameter sets on %d documents (= %d tasks) and %d processes'
                     % (n_params, n_docs, n_tasks, self.n_workers))
 
+        # distribute tasks to first "n_workers" workers
         logger.debug('distributing initial work')
         task_idx = 0
         for d, p in tasks[:self.n_workers]:
@@ -94,19 +132,22 @@ class MultiprocModelsRunner(object):
             self.tasks_queues[task_idx].put((d, p))
             task_idx += 1
 
+        # collect results for these workers
         worker_results = []
-        while task_idx < n_tasks:
+        while task_idx < n_tasks:  # until all tasks are distributed
             logger.debug('awaiting result')
             finished_worker, w_doc, w_params, w_result = self.results_queue.get()    # blocking
             logger.debug('> got result from worker %d' % finished_worker)
 
             worker_results.append((w_doc, w_params, w_result))
 
+            # the finished worker is free for a new task, send it
             d, p = tasks[task_idx]
             logger.debug('> sending task %d/%d to worker %d' % (task_idx + 1, n_tasks, finished_worker))
             self.tasks_queues[finished_worker].put((d, p))
             task_idx += 1
 
+        # collect results for the last submitted tasks
         logger.debug('awaiting final results')
         [q.join() for q in self.tasks_queues]   # block for last submitted tasks
 
@@ -116,8 +157,10 @@ class MultiprocModelsRunner(object):
 
         logger.info('multiproc models: finished')
 
+        # stop worker processes
         self.shutdown_workers()
 
+        # return results
         if self.got_named_data:
             res = defaultdict(list)
             for d, p, r in worker_results:
@@ -128,6 +171,7 @@ class MultiprocModelsRunner(object):
             return list(zip(p, r))
 
     def _setup_workers(self, worker_class):
+        """Create worker classes, their task queues and the result queue."""
         self.tasks_queues = []
         self.results_queue = mp.Queue()
         self.workers = []
@@ -141,10 +185,15 @@ class MultiprocModelsRunner(object):
             self.tasks_queues.append(task_q)
 
     def _new_worker(self, worker_class, i, task_queue, results_queue, data):
+        """Initiate new worker class."""
         return worker_class(i, task_queue, results_queue, data, name='%s#%d' % (str(worker_class), i))
 
     @staticmethod
     def _prepare_data(data):
+        """
+        Prepare 2D array/matrix `data` for parallel processing by converting it to COO sparse matrix (if necessary)
+        and creating shared data pointers for direct access for worker processes.
+        """
         if hasattr(data, 'dtype'):
             if not hasattr(data, 'shape') or len(data.shape) != 2:
                 raise ValueError('`data` must be a NumPy array/matrix or SciPy sparse matrix of two dimensions')
@@ -176,10 +225,30 @@ class MultiprocModelsRunner(object):
 
 
 class MultiprocModelsWorkerABC(mp.Process):
+    """
+    Abstract base class for parallel model computations worker class.
+    """
+
     package_name = None   # abstract. override in subclass
 
     def __init__(self, worker_id, tasks_queue, results_queue, data,
                  group=None, target=None, name=None, args=(), kwargs=None):
+        """
+        Initialize parallel model computations worker class with an ID `worker_id`, a queue to receive tasks from
+        `tasks_queue`, a queue to send results to `results_queue` and the `data` to operate on.
+
+        :param worker_id: process ID
+        :param tasks_queue: queue to receive tasks from
+        :param results_queue: queue to send results to
+        :param data: data to operate on; a dict mapping dataset label to a dataset; can be anything but is usually a
+                     tuple of shared data pointers for sparse matrix in COO format (see
+                     :meth:`tmtoolkit.topicmod.parallel.MultiprocModelsRunner._prepare_data`)
+        :param group: see Python's :class:`multiprocessing.Process` class
+        :param target: see Python's :class:`multiprocessing.Process` class
+        :param name: see Python's :class:`multiprocessing.Process` class
+        :param args: see Python's :class:`multiprocessing.Process` class
+        :param kwargs: see Python's :class:`multiprocessing.Process` class
+        """
         super(MultiprocModelsWorkerABC, self).__init__(group, target, name, args, kwargs or {})
 
         logger.debug('worker `%s`: creating worker with ID %d' % (self.name, worker_id))
@@ -187,6 +256,7 @@ class MultiprocModelsWorkerABC(mp.Process):
         self.tasks_queue = tasks_queue
         self.results_queue = results_queue
 
+        # set up data to operate on
         self.data_per_doc = {}
         for doc_label, mem in data.items():
             if isinstance(mem, tuple) and len(mem) == 3:
@@ -200,6 +270,10 @@ class MultiprocModelsWorkerABC(mp.Process):
                 self.data_per_doc[doc_label] = mem
 
     def run(self):
+        """
+        Run the process worker: Calls :meth:`~tmtoolkit.topicmod.parallel.MultiprocModelsWorkerABC.fit_model` on each
+        dataset and parameter set coming from the tasks queue.
+        """
         logger.debug('worker `%s`: run' % self.name)
 
         for doc, params in iter(self.tasks_queue.get, None):
@@ -216,9 +290,23 @@ class MultiprocModelsWorkerABC(mp.Process):
         self.tasks_queue.task_done()
 
     def fit_model(self, data, params):
+        """
+        Method stub to implement actually model fitting for `data` with parameter set `params`.
+
+        :param data: data passed to the model fitting algorithm
+        :param params: parameter set dict
+        :return: model fitting / evaluation results
+        """
         raise NotImplementedError('abstract base class method `fit_model` needs to be defined')
 
     def send_results(self, doc, params, results):
+        """
+        Put the results into the results queue.
+
+        :param doc: "document" / dataset label
+        :param params: used parameter set
+        :param results: generated results, e.g. fit model and/or evaluation results
+        """
         self.results_queue.put((self.worker_id, doc, params, results))
 
 
@@ -226,8 +314,29 @@ class MultiprocModelsWorkerABC(mp.Process):
 
 
 class MultiprocEvaluationRunner(MultiprocModelsRunner):
+    """
+    Specialization of :class:`~tmtoolkit.topicmod.parallel.MultiprocModelsRunner` for parallel model evaluations.
+    """
+
     def __init__(self, worker_class, available_metrics, data, varying_parameters, constant_parameters=None,
-                 metric=None, metric_options=None, n_max_processes=None, return_models=False):  # , n_folds=0
+                 metric=None, metric_options=None, n_max_processes=None, return_models=False):
+        """
+        Initialize evaluation runner.
+
+        :param worker_class: model computation worker class derived from
+                             :class:`~tmtoolkit.topicmod.parallel.MultiprocModelsWorkerABC`
+        :param available_metrics: list/tuple with available metrics as strings
+        :param data: the data that the workers use for computations; 2D (sparse) array/matrix
+        :param varying_parameters: list of dicts with parameters; each parameter set will be used in a separate
+                                   computation
+        :param constant_parameters: dict with parameters that are the same for all parallel computations
+        :param metric: string or list of strings; if given, use only this metric(s) for evaluation; must be subset of
+                       `available_metrics`
+        :param metric_options: dict of options for metric used metric(s)
+        :param n_max_processes: maximum number of worker processes to spawn
+        :param return_models: if True, also return the computed models in the evaluation results
+        """
+
         if isinstance(data, dict):
             raise ValueError('`data` cannot be a dict for evaluation')
 
@@ -260,15 +369,40 @@ class MultiprocEvaluationRunner(MultiprocModelsRunner):
         self.return_models = return_models
 
     def _new_worker(self, worker_class, i, task_queue, results_queue, data):
+        """Initiate new worker class. Also pass metrics to be used during evaluation."""
         return worker_class(i, self.eval_metric, self.eval_metric_options, self.return_models,
                             task_queue, results_queue, data, name='%s#%d' % (str(worker_class), i))
 
 
 class MultiprocEvaluationWorkerABC(MultiprocModelsWorkerABC):
+    """
+    Specialization of :class:`~tmtoolkit.topicmod.parallel.MultiprocModelsWorkerABC` for parallel model evaluations.
+    """
+
     def __init__(self, worker_id,
                  eval_metric, eval_metric_options, return_models,
                  tasks_queue, results_queue, data,
                  group=None, target=None, name=None, args=(), kwargs=None):
+        """
+        Initialize parallel model evaluations worker class with an ID `worker_id`, a queue to receive tasks from
+        `tasks_queue`, a queue to send results to `results_queue` and the `data` to operate on. Use evaluation
+        metrics `eval_metric`.
+
+        :param worker_id: process ID
+        :param eval_metric: list/tuple of strings of evaluation metrics to use
+        :param eval_metric_options: dict of options for metric used metric(s)
+        :param tasks_queue: queue to receive tasks from
+        :param results_queue: queue to send results to
+        :param data: data to operate on; a dict mapping dataset label to a dataset; can be anything but is usually a
+                     tuple of shared data pointers for sparse matrix in COO format (see
+                     :meth:`tmtoolkit.topicmod.parallel.MultiprocModelsRunner._prepare_data`)
+        :param group: see Python's :class:`multiprocessing.Process` class
+        :param target: see Python's :class:`multiprocessing.Process` class
+        :param name: see Python's :class:`multiprocessing.Process` class
+        :param args: see Python's :class:`multiprocessing.Process` class
+        :param kwargs: see Python's :class:`multiprocessing.Process` class
+        """
+
         super(MultiprocEvaluationWorkerABC, self).__init__(worker_id,
                                                            tasks_queue, results_queue, data,
                                                            group, target, name, args, kwargs)
