@@ -1,10 +1,6 @@
 """
 Common statistics from bag-of-words (BoW) matrices.
-
-Markus Konrad <markus.konrad@wzb.eu>
 """
-
-import itertools
 
 import numpy as np
 import datatable as dt
@@ -22,8 +18,9 @@ def doc_lengths(dtm):
     """
     Return the length, i.e. number of terms for each document in document-term-matrix `dtm`.
     This corresponds to the row-wise sums in `dtm`.
-    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw terms counts.
-    :return: NumPy array of size N (number of docs) with integers indicating the number of terms per document.
+
+    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw terms counts
+    :return: NumPy array of size N (number of docs) with integers indicating the number of terms per document
     """
     if dtm.ndim != 2:
         raise ValueError('`dtm` must be a 2D array/matrix')
@@ -72,42 +69,45 @@ def get_codoc_frequencies(dtm, min_val=1, proportions=False):
     return codoc_frequencies(dtm, min_val=min_val, proportions=proportions)
 
 
+def word_cooccurrence(dtm, min_val=1, proportions=False):
+    """
+    Calculate the co-document frequency (aka word co-occurrence) matrix. Alias for
+    :func:`~tmtoolkit.bow.bow_stats.codoc_frequencies`.
+
+    .. seealso:: :func:`~tmtoolkit.bow.bow_stats.codoc_frequencies`
+    """
+    return codoc_frequencies(dtm, min_val=min_val, proportions=proportions)
+
+
 def codoc_frequencies(dtm, min_val=1, proportions=False):
     """
-    For each unique pair of words `w1, w2` in the vocab of `dtm` (i.e. its columns), return how often both occur
-    together at least `min_val` times. If `proportions` is True, return proportions scaled to the number of documents
-    instead of absolute numbers.
+    Calculate the co-document frequency (aka word co-occurrence) matrix for a document-term matrix `dtm`, i.e. how often
+    each pair of tokens occurs together at least `min_val` times in the same document. If `proportions` is True,
+    return proportions scaled to the number of documents instead of absolute numbers.
 
+    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
     :param min_val: threshold for counting occurrences
     :param proportions: If `proportions` is True, return proportions scaled to the number of documents instead of
                         absolute numbers.
-    :return: NumPy array of size M (vocab size) indicating how often each unique pair of words `w1, w2` occurs at
-             least `min_val` times.
+    :return: co-document frequency (aka word co-occurrence) matrix with shape (vocab size, vocab size)
     """
     if dtm.ndim != 2:
         raise ValueError('`dtm` must be a 2D array/matrix')
 
-    n_docs, n_vocab = dtm.shape
-    if n_vocab < 2:
-        raise ValueError('`dtm` must have at least two columns (i.e. 2 unique words)')
+    if dtm.shape[1] < 2:
+        raise ValueError('`dtm` must have at least two columns')
 
-    word_in_doc = dtm >= min_val
+    if issparse(dtm) and dtm.format != 'csc':
+        dtm = dtm.tocsc()
 
-    codoc_freq = {}
-    for w1, w2 in itertools.combinations(range(n_vocab), 2):
-        if issparse(dtm):
-            w1_in_docs = word_in_doc[:, w1].A.flatten()
-            w2_in_docs = word_in_doc[:, w2].A.flatten()
-        else:
-            w1_in_docs = word_in_doc[:, w1]
-            w2_in_docs = word_in_doc[:, w2]
+    bin_dtm = (dtm >= min_val).astype(np.int)
 
-        freq = np.sum(w1_in_docs & w2_in_docs)
-        if proportions:
-            freq /= n_docs
-        codoc_freq[(w1, w2)] = freq
+    cooc = bin_dtm.T @ bin_dtm
 
-    return codoc_freq
+    if proportions:
+        return cooc / dtm.shape[0]
+    else:
+        return cooc
 
 
 @deprecated(deprecated_in='0.9.0', removed_in='0.10.0',
@@ -120,6 +120,7 @@ def term_frequencies(dtm, proportions=False):
     """
     Return the number of occurrences of each term in the vocab across all documents in document-term-matrix `dtm`.
     This corresponds to the column-wise sums in `dtm`.
+
     :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
     :param proportions: If `proportions` is True, return proportions scaled to the number of terms in the whole `dtm`.
     :return: NumPy array of size M (vocab size) with integers indicating the number of occurrences of each term in the
@@ -149,8 +150,9 @@ def get_term_proportions(dtm):
 
 def tf_binary(dtm):
     """
-    Transform raw count document-term-matrix `dtm` to binary term frequency matrix. This matrix contains 1 whereever
+    Transform raw count document-term-matrix `dtm` to binary term frequency matrix. This matrix contains 1 whenever
     a term occurred in a document, else 0.
+
     :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
     :return: (sparse) binary term frequency matrix of type integer of size NxM
     """
@@ -164,8 +166,10 @@ def tf_proportions(dtm):
     """
     Transform raw count document-term-matrix `dtm` to term frequency matrix with proportions, i.e. term counts
     normalized by document length.
+
     Note that this may introduce NaN values due to division by zero when a document is of length 0.
-    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
+
+    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts
     :return: (sparse) term frequency matrix of size NxM with proportions, i.e. term counts normalized by document length
     """
     if dtm.ndim != 2:
@@ -186,9 +190,10 @@ def tf_proportions(dtm):
 
 def tf_log(dtm, log_fn=np.log1p):
     """
-    Transform raw count document-term-matrix `dtm` to log-normalized term frequency matrix `log_fn(1 + dtm)`.
+    Transform raw count document-term-matrix `dtm` to log-normalized term frequency matrix ``log_fn(dtm)``.
+
     :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
-    :param log_fn: log function to use. default is NumPy's `log1p`, which calculates `log(1 + x)`.
+    :param log_fn: log function to use; default is NumPy's :func:`numpy.log1p`, which calculates ``log(1 + x)``
     :return: (sparse) log-normalized term frequency matrix of size NxM
     """
     if dtm.ndim != 2:
@@ -209,10 +214,12 @@ def tf_log(dtm, log_fn=np.log1p):
 def tf_double_norm(dtm, K=0.5):
     """
     Transform raw count document-term-matrix `dtm` to double-normalized term frequency matrix
-    `K + (1-K) * dtm / max{t in doc}`, where `max{t in doc}` is vector of size N containing the maximum term count per
-    document.
+    ``K + (1-K) * dtm / max{t in doc}``, where ``max{t in doc}`` is vector of size N containing the maximum term count
+    per document.
+
     Note that this may introduce NaN values due to division by zero when a document is of length 0.
-    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
+
+    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts
     :param K: normalization factor
     :return: double-normalized term frequency matrix of size NxM
     """
@@ -230,9 +237,9 @@ def tf_double_norm(dtm, K=0.5):
 def idf(dtm, smooth_log=1, smooth_df=1):
     """
     Calculate inverse document frequency (idf) vector from raw count document-term-matrix `dtm` with formula
-    `log(smooth_log + N / (smooth_df + df))`, where `N` is the number of documents, `df` is the document frequency
-    (see function `doc_frequencies()`) and `smooth_*` are smoothing constants. With default arguments, the formula
-    is thus `log(1 + N/(1+df))`.
+    ``log(smooth_log + N / (smooth_df + df))``, where ``N`` is the number of documents, ``df`` is the document frequency
+    (see function :func:`~tmtoolkit.bow.bow_stats.doc_frequencies`), `smooth_log` and `smooth_df` are smoothing
+    constants. With default arguments, the formula is thus ``log(1 + N/(1+df))``.
 
     Note that this may introduce NaN values due to division by zero when a document is of length 0.
 
@@ -257,8 +264,8 @@ def idf(dtm, smooth_log=1, smooth_df=1):
 def idf_probabilistic(dtm, smooth=1):
     """
     Calculate probabilistic inverse document frequency (idf) vector from raw count document-term-matrix `dtm` with
-    formula `log(smooth + (N - df) / df)`, where `N` is the number of documents and `df` is the document frequency (see
-    function `doc_frequencies()`).
+    formula ``log(smooth + (N - df) / df)``, where ``N`` is the number of documents and ``df`` is the document
+    frequency (see function :func:`~tmtoolkit.bow.bow_stats.doc_frequencies`).
 
     :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
     :param smooth: smoothing constant (setting this to 0 can lead to -inf results)
@@ -280,12 +287,12 @@ def idf_probabilistic(dtm, smooth=1):
 def tfidf(dtm, tf_func=tf_proportions, idf_func=idf, **kwargs):
     """
     Calculate tfidf (term frequency inverse document frequency) matrix from raw count document-term-matrix `dtm` with
-    matrix multiplication `tf * diag(idf)`, where `tf` is the term frequency matrix `tf_funct(dtm)` and `idf` is the
-    document frequency vector `idf_func(dtm)`.
+    matrix multiplication ``tf * diag(idf)``, where `tf` is the term frequency matrix ``tf_func(dtm)`` and ``idf`` is
+    the document frequency vector ``idf_func(dtm)``.
 
-    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts.
-    :param tf_func: function to calculate term-frequency matrix. see `tf_*` functions in this module.
-    :param idf_func: function to calculate inverse document frequency vector. see `tf_*` functions in this module.
+    :param dtm: (sparse) document-term-matrix of size NxM (N docs, M is vocab size) with raw term counts
+    :param tf_func: function to calculate term-frequency matrix; see ``tf_*`` functions in this module
+    :param idf_func: function to calculate inverse document frequency vector; see ``tf_*`` functions in this module
     :param kwargs: additional parameters passed to `tf_func` or `idf_func` like `K` or `smooth` (depending on which
                    parameters these functions except)
     :return: (sparse) tfidf matrix of size NxM
@@ -317,16 +324,17 @@ def tfidf(dtm, tf_func=tf_proportions, idf_func=idf, **kwargs):
         return tf_mat * idf_vec
 
 
-def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=False, data_table_doc_labels=None):
+def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=False, datatable_doc_labels=None):
     """
     For each row (i.e. document) in a (sparse) document-term-matrix `mat`, do the following:
+
     1. filter all values according to `lo_thresh` and `hi_thresh`
     2. sort values and the corresponding terms from `vocab` according to `ascending`
     3. optionally select the top `top_n` terms
     4. generate a list with pairs of terms and values
 
     Return the collected lists for each row or convert the result to a data frame if document labels are passed via
-    `data_frame_doc_labels` (see shortcut function `sorted_terms_data_table`).
+    `data_frame_doc_labels` (see shortcut function :func:`~tmtoolkit.bow.bow_stats.sorted_terms_datatable`).
 
     :param mat: (sparse) document-term-matrix `mat` (may be tf-idf transformed or any other transformation)
     :param vocab: list or array of vocabulary corresponding to columns in `mat`
@@ -334,7 +342,7 @@ def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=F
     :param hi_tresh: if not None, filter for values lesser than or equal `hi_thresh`
     :param top_n: if not None, select only the top `top_n` terms
     :param ascending: sorting direction
-    :param data_table_doc_labels: optional list/array of document labels corresponding to `mat` rows
+    :param datatable_doc_labels: optional list/array of document labels corresponding to `mat` rows
     :return: list of list with tuples (term, value) or data table with columns "doc", "term", "value"
              if `data_frame_doc_labels` is given
     """
@@ -349,7 +357,7 @@ def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=F
     if top_n is not None and top_n < 1:
         raise ValueError('`top_n` must be at least 1')
 
-    if data_table_doc_labels is not None and len(data_table_doc_labels) != mat.shape[0]:
+    if datatable_doc_labels is not None and len(datatable_doc_labels) != mat.shape[0]:
         raise ValueError('length of `data_frame_doc_labels` must match number of rows in `mat`')
 
     if not isinstance(vocab, np.ndarray):
@@ -401,15 +409,15 @@ def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=F
         rowsize = len(row_terms)
         assert rowsize == len(row_vals)
 
-        if data_table_doc_labels:
+        if datatable_doc_labels is not None:
             if rowsize > 0:
-                res.append(dt.Frame({'doc': np.repeat(data_table_doc_labels[i], repeats=rowsize),
+                res.append(dt.Frame({'doc': np.repeat(datatable_doc_labels[i], repeats=rowsize),
                                      'token': row_terms,
                                      'value': row_vals}))
         else:
             res.append(list(zip(row_terms, row_vals)))
 
-    if data_table_doc_labels:
+    if datatable_doc_labels is not None:
         if res:
             return dt.rbind(*res)
         else:
@@ -418,9 +426,9 @@ def sorted_terms(mat, vocab, lo_thresh=0, hi_tresh=None, top_n=None, ascending=F
         return res
 
 
-def sorted_terms_data_table(mat, vocab, doc_labels, lo_thresh=0, hi_tresh=None, top_n=None, ascending=False):
+def sorted_terms_datatable(mat, vocab, doc_labels, lo_thresh=0, hi_tresh=None, top_n=None, ascending=False):
     """
-    Shortcut function for `sorted_terms` which generates a data table with `doc_labels`.
+    Shortcut function for :func:`~tmtoolkit.bow.bow_stats.sorted_terms` which generates a data table with `doc_labels`.
 
     :param mat: (sparse) document-term-matrix `mat` (may be tf-idf transformed or any other transformation)
     :param vocab: list or array of vocabulary corresponding to columns in `mat`
@@ -432,4 +440,4 @@ def sorted_terms_data_table(mat, vocab, doc_labels, lo_thresh=0, hi_tresh=None, 
     :return: data table with columns "doc", "term", "value"
     """
     return sorted_terms(mat, vocab, lo_thresh=lo_thresh, hi_tresh=hi_tresh, top_n=top_n,
-                        ascending=ascending, data_table_doc_labels=doc_labels)
+                        ascending=ascending, datatable_doc_labels=doc_labels)
