@@ -1,5 +1,11 @@
 """
-Parallel text processing with `TMPreproc` class.
+Parallel text processing with :class:`TMPreproc` class.
+
+Implements several methods for text processing similar to the functional API but processing is done in parallel on a
+multi-core machine. For large amounts of text, using TMPreproc on a multi-core machine will be much faster than
+sequential processing with the functional API.
+
+Markus Konrad <markus.konrad@wzb.eu>
 """
 
 import string
@@ -30,11 +36,35 @@ logger.addHandler(logging.NullHandler())
 
 
 class TMPreproc:
-    def __init__(self, docs, language=defaults.language, n_max_processes=None,
+    """
+    TMPreproc implements a class for parallel text processing. The API implements a state machine, i.e. you create a
+    TMPreproc instance with text documents and modify them by calling methods like "to_lowercase", etc.
+    """
+
+    def __init__(self, docs, language=None, n_max_processes=None,
                  stopwords=None, punctuation=None, special_chars=None,
                  tokenizer=None, pos_tagger=None, pos_tagset=None, stemmer=None, lemmatizer=None):
-        if isinstance(docs, Corpus):
-            docs = docs.docs
+        """
+        Create a parallel text processing instance by passing a dictionary of raw texts `docs` with document label
+        to document text mapping. You can pass a :class:`~tmtoolkit.corpus.Corpus` instance because it implements the
+        dictionary methods.
+
+        TMPreproc will start `n_max_processes` sub-processes and distribute the documents on them for parallel
+        processing.
+
+        :param docs: documents dictionary ("corpus") with document label to document text mapping
+        :param language: documents language used for language-dependent methods such as POS tagging or lemmatization
+        :param n_max_processes: max. number of sub-processes for parallel processing; uses the number of CPUs on the
+                                current machine if None is passed
+        :param stopwords: provide manual stopword list or use default stopword list for given language
+        :param punctuation: provide manual punctuation symbols list or use default list from :func:`string.punctuation`
+        :param special_chars: provide manual special characters list or use default list from :func:`string.punctuation`
+        :param tokenizer: provide custom tokenizer function or use :func:`~tmpreproc.preprocess.tokenize`
+        :param pos_tagger: provide custom POS tagger function or use :func:`~tmpreproc.preprocess.pos_tag`
+        :param pos_tagset: custom tagset label for custom POS tagger
+        :param stemmer: provide custom stemmer function or use :func:`~tmpreproc.preprocess.stem`
+        :param lemmatizer: provide custom lemmatizer function or use :func:`~tmpreproc.preprocess.lemmatize`
+        """
 
         if docs is not None:
             require_dictlike(docs)
@@ -59,45 +89,48 @@ class TMPreproc:
         self._cur_vocab_counts = None
         self._cur_dtm = None
 
-        self.language = language   # document language
+        self.language = language or defaults.language   # document language
         self.ngrams_as_tokens = False
 
         if stopwords is None:      # load default stopword list for this language
             import nltk
-            self.stopwords = nltk.corpus.stopwords.words(language)
+            self.stopwords = nltk.corpus.stopwords.words(self.language)
         else:                      # set passed stopword list
+            require_listlike(stopwords)
             self.stopwords = stopwords
 
         if punctuation is None:    # load default punctuation list
             self.punctuation = list(string.punctuation)
         else:
+            require_listlike(punctuation)
             self.punctuation = punctuation
 
         if special_chars is None:
             self.special_chars = list(string.punctuation)
         else:
+            require_listlike(special_chars)
             self.special_chars = special_chars
 
         if tokenizer is None:
-            self.tokenizer = partial(tokenize, language=language)
+            self.tokenizer = partial(tokenize, language=self.language)
         else:
             self.tokenizer = tokenizer
 
         if pos_tagger is None:
-            tagger_instance, self.pos_tagset = load_pos_tagger_for_language(language)
-            self.pos_tagger = partial(pos_tag, tagger_instance=tagger_instance)
+            tagger_instance, self.pos_tagset = load_pos_tagger_for_language(self.language)
+            self.pos_tagger = partial(pos_tag, tagger_instance=tagger_instance, doc_meta_key='meta_pos')
         else:
             self.pos_tagger = pos_tagger
             self.pos_tagset = pos_tagset
 
         if stemmer is None:
             import nltk
-            self.stemmer = partial(stem, stemmer_instance=nltk.stem.SnowballStemmer(language))
+            self.stemmer = partial(stem, stemmer_instance=nltk.stem.SnowballStemmer(self.language))
         else:
             self.stemmer = stemmer
 
         if lemmatizer is None:
-            self.lemmatizer = partial(lemmatize, lemmatizer_fn=load_lemmatizer_for_language(language))
+            self.lemmatizer = partial(lemmatize, lemmatizer_fn=load_lemmatizer_for_language(self.language))
         else:
             self.lemmatizer = lemmatizer
 
@@ -118,14 +151,22 @@ class TMPreproc:
 
     @property
     def n_docs(self):
+        """Number of documents."""
         return len(self.doc_labels)
 
     @property
+    def vocabulary_size(self):
+        """Number of unique tokens across all documents."""
+        return len(self.vocabulary)
+
+    @property
     def n_tokens(self):
+        """Number of tokens in all documents (sum of document lengths)."""
         return sum(self.doc_lengths.values())
 
     @property
     def doc_labels(self):
+        """Document labels as sorted list."""
         if self._cur_doc_labels is None:
             self._cur_doc_labels = sorted(flatten_list(self._get_results_seq_from_workers('get_doc_labels')))
 
@@ -133,15 +174,18 @@ class TMPreproc:
 
     @property
     def doc_lengths(self):
+        """Document lengths as dict with mapping document label to document length (number of tokens in doc.)."""
         return dict(zip(self.tokens.keys(), doc_lengths(self.tokens.values())))
 
     @property
     def tokens(self):
+        """Document tokens as dict with mapping document label to list of tokens."""
         return self.get_tokens(with_metadata=False)
 
     @property
     def tokens_with_metadata(self):
-        return self.get_tokens(with_metadata=True, as_data_tables=True)
+        """Document tokens with metadata (e.g. POS tag) as dict with mapping document label to datatable."""
+        return self.get_tokens(with_metadata=True, as_datatables=True)
 
     @property
     def tokens_datatable(self):
@@ -149,7 +193,7 @@ class TMPreproc:
         Tokens and metadata as datatable with columns "doc" (document label), "position" (token position in
         the document), "token" and optional meta data columns.
         """
-        tokens = self.get_tokens(non_empty=True, with_metadata=True, as_data_tables=True)
+        tokens = self.get_tokens(non_empty=True, with_metadata=True, as_datatables=True)
         dfs = []
         for dl, df in tokens.items():
             n = df.shape[0]
@@ -179,16 +223,25 @@ class TMPreproc:
 
     @property
     def tokens_with_pos_tags(self):
+        """
+        Document tokens with POS tag as dict with mapping document label to datatable. The datatables have two
+        columns, ``token`` and ``meta_pos``.
+        """
         self._require_pos_tags()
         return {dl: df[:, ['token', 'meta_pos']]
-                for dl, df in self.get_tokens(with_metadata=True, as_data_tables=True).items()}
+                for dl, df in self.get_tokens(with_metadata=True, as_datatables=True).items()}
 
     @property
     def vocabulary(self):
+        """Corpus vocabulary, i.e. sorted list of all tokens that occur across all documents."""
         return self.get_vocabulary()
 
     @property
     def vocabulary_counts(self):
+        """
+        :class:`collections.Counter()` instance of vocabulary containing counts of occurrences of tokens across all
+        documents.
+        """
         if self._cur_vocab_counts is not None:
             return self._cur_vocab_counts
 
@@ -202,36 +255,82 @@ class TMPreproc:
 
     @property
     def vocabulary_abs_doc_frequency(self):
+        """
+        Absolute document frequency per vocabulary token as dict with token to document frequency mapping.
+        Document frequency is the measure of how often a token occurs *at least once* in a document.
+        Example:
+        ::
+
+            doc tokens
+            --- ------
+            A   z, z, w, x
+            B   y, z, y
+            C   z, z, y, z
+
+            document frequency df(z) = 3  (occurs in all 3 documents)
+            df(x) = df(w) = 1 (occur only in A)
+            df(y) = 2 (occurs in B and C)
+            ...
+        """
         return self._workers_vocab_doc_frequencies
 
     @property
     def vocabulary_rel_doc_frequency(self):
+        """
+        Same as :attr:`~TMPreproc.vocabulary_abs_doc_frequency` but normalized by number of documents,
+        i.e. relative document frequency.
+        """
         return {w: n/self.n_docs for w, n in self._workers_vocab_doc_frequencies.items()}
 
     @property
     def ngrams_generated(self):
+        """Indicates if n-grams were generated before (True if yes, else False)."""
         return len(self.ngrams) > 0
 
     @property
     def ngrams(self):
+        """
+        Generated n-grams as dict with mapping document label to list of n-grams. Each list of n-grams (i.e. each
+        document) in turn contains lists of size ``n`` (i.e. two if you generated bigrams).
+        """
         return self.get_ngrams()
 
     @property
     def pos_tagged(self):
+        """Indicates if documents were POS tagged. True if yes, otherwise False."""
         meta_keys = self.get_available_metadata_keys()
         return 'pos' in meta_keys
 
     @property
     def dtm(self):
+        """
+        Generate and return a sparse document-term matrix of shape ``(n_docs, n_vocab)`` where ``n_docs`` is the number
+        of documents and ``n_vocab`` is the vocabulary size.
+        """
         return self.get_dtm()
 
     def shutdown_workers(self):
+        """
+        Manually send the shutdown signal to all worker processes.
+
+        Normally you don't need to call this manually as the worker processes are killed automatically when the
+        TMPreproc instance is removed. However, if you need to free resources immediately, you can use this method
+        as it is also used in the tests.
+        """
         try:   # may cause exception when the logger is actually already destroyed
             logger.info('sending shutdown signal to workers')
         except: pass
         self._send_task_to_workers(None)
 
     def save_state(self, picklefile):
+        """
+        Save the current state of this TMPreproc instance to disk, i.e. to the pickle file `picklefile`.
+        The state can be restored from this file using :meth:`~TMPreproc.load_state` or class method
+        :meth:`~TMPreproc.from_state`.
+
+        :param picklefile: disk file to store the state to
+        :return: this instance
+        """
         # attributes for this instance ("manager instance")
         logger.info('saving state to file `%s`' % picklefile)
 
@@ -241,6 +340,14 @@ class TMPreproc:
         return self
 
     def load_state(self, file_or_stateobj):
+        """
+        Restore a state by loading either a pickled state from disk as saved with :meth:`~TMPreproc.save_state` or by
+        loading a state object directly.
+
+        :param file_or_stateobj: either path to a pickled file as saved with :meth:`~TMPreproc.save_state` or a
+                                 state object
+        :return: this instance as restored from the passed file / object
+        """
         if isinstance(file_or_stateobj, str):
             logger.info('loading state from file `%s`' % file_or_stateobj)
             state_data = unpickle_file(file_or_stateobj)
@@ -267,9 +374,13 @@ class TMPreproc:
 
     def load_tokens(self, tokens):
         """
-        Load tokens `tokens` into TMPreproc in the same format as they are returned by `self.tokens` or
-        `self.tokens_with_metadata`, i.e. as dict with mapping: document label -> document tokens array or
-        document data frame.
+        Load tokens `tokens` into :class:`TMPreproc` in the same format as they are returned by
+        :attr:`~TMPreproc.tokens` or :attr:`~TMPreproc.tokens_with_metadata`, i.e. as dict with mapping:
+        document label -> document tokens array or document data frame.
+
+        :param tokens: dict of tokens as returned by :attr:`~TMPreproc.tokens` or
+                       :attr:`~TMPreproc.tokens_with_metadata`
+        :return: this instance
         """
         require_dictlike(tokens)
 
@@ -298,9 +409,12 @@ class TMPreproc:
 
     def load_tokens_datatable(self, tokendf):
         """
-        Load tokens dataframe `tokendf` into TMPreproc in the same format as they are returned by `self.tokens_frame`,
-        i.e. as data frame with hierarchical indices "doc" and "position" and at least a column "token" plus optional
-        columns like "meta_pos", etc.
+        Load tokens dataframe `tokendf` into :class:`TMPreproc` in the same format as they are returned by
+        :attr:`~TMPreproc.tokens_datatable`, i.e. as data frame with indices "doc" and "position" and
+        at least a column "token" plus optional columns like "meta_pos", etc.
+
+        :param tokendf: tokens datatable Frame object as returned by :attr:`~TMPreproc.tokens_datatable`
+        :return: this instance
         """
         if not isinstance(tokendf, dt.Frame):
             raise ValueError('`tokendf` must be a datatable Frame object')
@@ -322,6 +436,8 @@ class TMPreproc:
         """
         Copy a TMPreproc object including all its present state (tokens, meta data, etc.).
         Performs a deep copy.
+
+        :return: deep copy of the current TMPreproc instance
         """
         return self.copy()
 
@@ -329,18 +445,31 @@ class TMPreproc:
         """
         Copy a TMPreproc object including all its present state (tokens, meta data, etc.).
         Performs a deep copy.
+
+        :return: deep copy of the current TMPreproc instance
         """
         return self.copy()
 
     def copy(self):
         """
-        Copy a TMPreproc object including all its present state (tokens, meta data, etc.).
+        Copy a TMPreproc instance including all its present state (tokens, meta data, etc.).
         Performs a deep copy.
+
+        :return: deep copy of the current TMPreproc instance
         """
         return TMPreproc.from_state(self._create_state_object(deepcopy_attrs=True))
 
     @classmethod
     def from_state(cls, file_or_stateobj, **init_kwargs):
+        """
+        Create a new TMPreproc instance by loading a state from by loading either a pickled state from disk as saved
+        with :meth:`~TMPreproc.save_state` or by loading a state object directly.
+
+        :param file_or_stateobj: either path to a pickled file as saved with :meth:`~TMPreproc.save_state` or a
+                                 state object
+        :param init_kwargs: dict of arguments passed to :meth:`~TMPreproc.__init__`
+        :return: new instance as restored from the passed file / object
+        """
         if 'docs' in init_kwargs.keys():
             raise ValueError('`docs` cannot be passed as argument when loading state')
         init_kwargs['docs'] = None
@@ -349,6 +478,16 @@ class TMPreproc:
 
     @classmethod
     def from_tokens(cls, tokens, **init_kwargs):
+        """
+        Create a new TMPreproc instance by loading `tokens` in the same format as they are returned by
+        :attr:`~TMPreproc.tokens` or :attr:`~TMPreproc.tokens_with_metadata`, i.e. as dict with mapping: document
+        label -> document tokens array or document data frame.
+
+        :param tokens: dict of tokens as returned by :attr:`~TMPreproc.tokens` o
+                       :attr:`~TMPreproc.tokens_with_metadata`
+        :param init_kwargs: dict of arguments passed to :meth:`~TMPreproc.__init__`
+        :return: new instance with passed tokens
+        """
         if 'docs' in init_kwargs.keys():
             raise ValueError('`docs` cannot be passed as argument when loading tokens')
         init_kwargs['docs'] = None
@@ -357,6 +496,15 @@ class TMPreproc:
 
     @classmethod
     def from_tokens_datatable(cls, tokensdf, **init_kwargs):
+        """
+        Create a new TMPreproc instance by loading tokens dataframe `tokendf` in the same format as it is returned by
+        :meth:`~TMPreproc.tokens_datatable`, i.e. as data frame with hierarchical indices "doc" and "position" and at
+        least a column "token" plus optional columns like "meta_pos", etc.
+
+        :param tokendf: tokens datatable Frame object as returned by :meth:`~TMPreproc.tokens_datatable`
+        :param init_kwargs: dict of arguments passed to :meth:`~TMPreproc.__init__`
+        :return: new instance with passed tokens
+        """
         if 'docs' in init_kwargs.keys():
             raise ValueError('`docs` cannot be passed as argument when loading a token datatable')
         init_kwargs['docs'] = None
@@ -369,14 +517,27 @@ class TMPreproc:
     def tokenize(self):
         return self
 
-    def get_tokens(self, non_empty=False, with_metadata=True, as_data_tables=False):
+    def get_tokens(self, non_empty=False, with_metadata=True, as_datatables=False):
+        """
+        Return document tokens as dict with mapping document labels to document tokens. The format of the tokens
+        depends on the passed arguments: If `as_datatables` is True, each document is a datatable with at least
+        the column ``"token"`` and optional ``"meta_..."`` columns (e.g. ``"meta_pos"`` for POS tags) if
+        `with_metadata` is True.
+        If `as_datatables` is False, the result documents are either plain lists of tokens if `with_metadata` is
+        False, or they're dicts of lists with keys ``"token"`` and optional ``"meta_..."`` keys.
+
+        :param non_empty: remove empty documents from the result set
+        :param with_metadata: add meta data to results (e.g. POS tags)
+        :param as_datatables: return results as datatables
+        :return: dict mapping document labels to document tokens
+        """
         tokens = self._workers_tokens
         meta_keys = self.get_available_metadata_keys()
 
         if not with_metadata:  # doc label -> token array
             tokens = {dl: doc['token'] for dl, doc in tokens.items()}
 
-        if as_data_tables:
+        if as_datatables:
             if with_metadata:  # doc label -> doc data frame with token and meta data columns
                 tokens_dfs = {}
                 for dl, doc in tokens.items():
@@ -397,11 +558,11 @@ class TMPreproc:
             return tokens
 
     def get_kwic(self, search_token, context_size=2, match_type='exact', ignore_case=False, glob_method='match',
-                 inverse=False, with_metadata=False, as_data_table=False, non_empty=False, glue=None,
+                 inverse=False, with_metadata=False, as_datatable=False, non_empty=False, glue=None,
                  highlight_keyword=None):
         """
         Perform keyword-in-context (kwic) search for `search_token`. Uses similar search parameters as
-        `filter_tokens()`.
+        :meth:`~TMPreproc.filter_tokens`.
 
         :param search_token: search pattern
         :param context_size: either scalar int or tuple (left, right) -- number of surrounding words in keyword context.
@@ -414,7 +575,7 @@ class TMPreproc:
                             behavior as Python's `re.match` or `re.search`).
         :param inverse: Invert the matching results.
         :param with_metadata: Also return metadata (like POS) along with each token.
-        :param as_data_table: Return result as data frame with indices "doc" (document label) and "context" (context
+        :param as_datatable: Return result as data frame with indices "doc" (document label) and "context" (context
                               ID per document) and optionally "position" (original token position in the document) if
                               tokens are not glued via `glue` parameter.
         :param non_empty: If True, only return non-empty result documents.
@@ -422,7 +583,7 @@ class TMPreproc:
         :param highlight_keyword: If not None, this must be a string which is used to indicate the start and end of the
                                   matched keyword.
         :return: Return dict with `document label -> kwic for document` mapping or a data frame, depending
-        on `as_data_table`.
+                 on `as_datatable`.
         """
         if isinstance(context_size, int):
             context_size = (context_size, context_size)
@@ -433,8 +594,8 @@ class TMPreproc:
             raise ValueError('if `highlight_keyword` is given, it must be of type str')
 
         if glue:
-            if with_metadata or as_data_table:
-                raise ValueError('when `glue` is set to True, `with_metadata` and `as_data_table` must be False')
+            if with_metadata or as_datatable:
+                raise ValueError('when `glue` is set to True, `with_metadata` and `as_datatable` must be False')
             if not isinstance(glue, str):
                 raise ValueError('if `glue` is given, it must be of type str')
 
@@ -444,7 +605,7 @@ class TMPreproc:
                                                           search_token=search_token,
                                                           highlight_keyword=highlight_keyword,
                                                           with_metadata=with_metadata,
-                                                          with_window_indices=as_data_table,
+                                                          with_window_indices=as_datatable,
                                                           match_type=match_type,
                                                           ignore_case=ignore_case,
                                                           glob_method=glob_method,
@@ -456,12 +617,13 @@ class TMPreproc:
             kwic.update(worker_kwic)
 
         return _finalize_kwic_results(kwic, non_empty=non_empty, glue=glue,
-                                      as_data_table=as_data_table, with_metadata=with_metadata)
+                                      as_datatable=as_datatable, with_metadata=with_metadata)
 
     def get_kwic_table(self, search_token, context_size=2, match_type='exact', ignore_case=False, glob_method='match',
                        inverse=False, glue=' ', highlight_keyword='*'):
         """
-        Shortcut for `get_kwic` to directly return a data frame table with highlighted keywords in context.
+        Shortcut for :meth:`~TMPreproc.get_kwic` to directly return a data frame table with highlighted keywords in
+        context.
 
         :param search_token: search pattern
         :param context_size: either scalar int or tuple (left, right) -- number of surrounding words in keyword context.
@@ -482,7 +644,7 @@ class TMPreproc:
 
         kwic = self.get_kwic(search_token=search_token, context_size=context_size, match_type=match_type,
                              ignore_case=ignore_case, glob_method=glob_method, inverse=inverse,
-                             with_metadata=False, as_data_table=False, non_empty=True,
+                             with_metadata=False, as_datatable=False, non_empty=True,
                              glue=glue, highlight_keyword=highlight_keyword)
 
         return _datatable_from_kwic_results(kwic)
@@ -490,7 +652,8 @@ class TMPreproc:
     def glue_tokens(self, patterns, glue='_', match_type='exact', ignore_case=False, glob_method='match',
                     inverse=False):
         """
-        Match N *subsequent* tokens to the N patterns in `patterns` using match options like in `filter_tokens`.
+        Match N *subsequent* tokens to the N patterns in `patterns` using match options like in
+        :meth:`~TMPreproc.filter_tokens`.
         Join the matched tokens by glue string `glue`. Replace these tokens in the documents.
 
         If there is metadata, the respective entries for the joint tokens are set to None.
@@ -532,7 +695,10 @@ class TMPreproc:
 
     def get_vocabulary(self, sort=True):
         """
-        Return the vocabulary, i.e. the list of unique words across all documents, as sorted NumPy array.
+        Return the vocabulary, i.e. the list of unique words across all documents, as (sorted) list.
+
+        :param sort: the vocabulary alphabetically
+        :return: list of tokens in the vocabulary
         """
         tokens = set(flatten_list(self._get_results_seq_from_workers('get_vocab')))
 
@@ -542,43 +708,144 @@ class TMPreproc:
             return list(tokens)
 
     def get_ngrams(self, non_empty=False):
+        """
+        Return generated n-grams as dict with mapping document labels to document n-grams list. Each list of n-grams
+        (i.e. each  document) in turn contains lists of size ``n`` (i.e. two if you generated bigrams).
+
+        :param non_empty: remove empty documents from the result set
+        :return: dict mapping document labels to document n-grams list
+        """
         if non_empty:
             return {dl: dtok for dl, dtok in self._workers_ngrams.items() if len(dtok) > 0}
         else:
             return self._workers_ngrams
 
     def get_available_metadata_keys(self):
+        """
+        Return set of available meta data keys, e.g. "pos" for POS tags if :meth:`~TMPreproc.pos_tag` was called before.
+
+        :return: set of available meta data keys
+        """
         keys = self._get_results_seq_from_workers('get_available_metadata_keys')
 
         return set(flatten_list(keys))
 
     def add_stopwords(self, stopwords):
+        """
+        Add more stop words to the set of stop words used in :meth:`~TMPreproc.clean_tokens`.
+
+        :param stopwords: list, tuple or set or stop words
+        :return: this instance
+        """
         require_listlike_or_set(stopwords)
         self.stopwords += stopwords
 
         return self
 
     def add_punctuation(self, punctuation):
+        """
+        Add more characters to the set of punctuation characters used in :meth:`~TMPreproc.clean_tokens`.
+
+        :param punctuation: list, tuple or set of punctuation characters
+        :return: this instance
+        """
         require_listlike_or_set(punctuation)
         self.punctuation += punctuation
 
         return self
 
     def add_special_chars(self, special_chars):
+        """
+        Add more characters to the set of "special characters" used in
+        :meth:`~TMPreproc.remove_special_chars_in_tokens`.
+
+        :param special_chars: list, tuple or set of special characters
+        :return: this instance
+        """
         require_listlike_or_set(special_chars)
         self.special_chars += special_chars
 
         return self
 
     def add_metadata_per_token(self, key, data, default=None):
+        """
+        Add a meta data value per token match, where `key` is the meta data label and `data` is a dict that maps
+        tokens to the respective meta data values. If a token existing in this instance is not listed in `data`,
+        the value `default` is taken instead. Example::
+
+            preproc = TMPreproc(docs={'a': 'This is a test document.',
+                                      'b': 'This is another test, test, test.'})
+            preproc.tokens_datatable
+
+        Output (note that there's no meta data column for the tokens)::
+
+                doc  position  token
+            --  ---  --------  --------
+             0  a           0  This
+             1  a           1  is
+             2  a           2  a
+             3  a           3  test
+            [...]
+
+        Now we add a meta data with the key ``interesting``, e.g. indicating which tokens we deem interesting. For every
+        occurrence of the token "test", this should be set to True (1). All other tokens by default get False (0)::
+
+            preproc.add_metadata_per_token('interesting', {'test': True}, default=False)
+            preproc.tokens_datatable
+
+        New output with additional column ``meta_interesting``::
+
+                doc  position  token     meta_interesting
+            --  ---  --------  --------  ----------------
+             0  a           0  This                     0
+             1  a           1  is                       0
+             2  a           2  a                        0
+             3  a           3  test                     1
+
+
+        :param key: meta data key, i.e. label as string
+        :param data: dict that maps tokens to the respective meta data values
+        :param default: default meta data value for tokens that do not appear in `data`
+        :return: this instance
+        """
         self._add_metadata('add_metadata_per_token', key, data, default)
         return self
 
     def add_metadata_per_doc(self, key, data, default=None):
+        """
+        Add a list of meta data values per document, where `key` is the meta data label and `data` is a dict that
+        maps document labels to meta data values. The length of the values of each document must match the number of
+        tokens for the respective document. If a document that exists in this instance is not part of `data`,
+        the value `default` will be repeated ``len(document)`` times.
+
+        Suppose you three documents named a, b, c with respective document lengths (i.e. number of tokens) 5, 3, 6. You
+        want to add meta data labelled as ``token_category``. You can do so by passing a dict with lists of values for
+        each document:
+        ::
+
+            preproc.add_metadata_per_doc('token_category', {
+                'a': ['x', 'y', 'z', 'y', 'z'],
+                'b': ['z', 'y', 'z'],
+                'c': ['x', 'x', 'x', 'y', 'x', 'x'],
+            })
+
+        :param key: meta data key, i.e. label as string
+        :param data: dict that maps document labels to meta data values
+        :param default: default value for documents not listed in `data`
+        :return: this instance
+        """
         self._add_metadata('add_metadata_per_doc', key, data, default)
         return self
 
     def remove_metadata(self, key):
+        """
+        Remove meta data information previously added by :meth:`~TMPreproc.pos_tag` or
+        :meth:`~TMPreproc.add_metadata_per_token`/:meth:`~TMPreproc.add_metadata_per_doc`
+        and identified by meta data key `key`.
+
+        :param key: meta data key, i.e. label as string
+        :return: this instance
+        """
         self._invalidate_workers_tokens()
 
         logger.info('removing metadata key')
@@ -591,6 +858,16 @@ class TMPreproc:
         return self
 
     def generate_ngrams(self, n):
+        """
+        Generate n-grams of length `n`. They are then available in the :attr:`~TMPreproc.ngrams` property.
+        Use `join_ngrams` to convert them to normal tokens by joining them.
+
+        :param n: length of n-grams, must be >= 2
+        :return: this instance
+        """
+        if n < 2:
+            raise ValueError("`n` must be at least 2")
+
         self._invalidate_workers_ngrams()
 
         logger.info('generating ngrams')
@@ -598,11 +875,15 @@ class TMPreproc:
 
         return self
 
-    def use_joined_ngrams_as_tokens(self, join_str=' '):
+    def join_ngrams(self, join_str=' '):
         """
         Use the generated n-grams as tokens by joining them via `join_str`. After this operation, the joined n-grams
-        are available as `.tokens` but the original n-grams will be removed and `.ngrams_generated` is reset to False.
-        Requires that n-grams have been generated with `.generate_ngrams()` before.
+        are available as :attr:`~TMPreproc.tokens` but the original n-grams will be removed and
+        :attr:`~TMPreproc.ngrams_generated` is reset to False.
+        Requires that n-grams have been generated with :meth:`~TMPreproc.generate_ngrams` before.
+
+        :param join_str: string use to "glue" the grams
+        :return: this instance
         """
         self._require_ngrams()
         self._invalidate_workers_tokens()
@@ -615,6 +896,17 @@ class TMPreproc:
         return self
 
     def transform_tokens(self, transform_fn):
+        """
+        Transform tokens in all documents by applying `transform_fn` to each document's tokens individually.
+
+        If `transform_fn` is "pickable" (e.g. ``pickle.dumps(transform_fn)`` doesn't raise an exception), the
+        function is applied in parallel. If not, the function is applied to the documents sequentially, which may
+        be very slow.
+
+        :param transform_fn: a function to apply to all documents' tokens; it must accept a single token string and
+                             vice-versa return single token string
+        :return: this instance
+        """
         if not callable(transform_fn):
             raise ValueError('`transform_fn` must be callable')
 
@@ -648,6 +940,11 @@ class TMPreproc:
         return self
 
     def tokens_to_lowercase(self):
+        """
+        Convert all tokens to lower-case form.
+
+        :return: this instance
+        """
         self._invalidate_workers_tokens()
 
         logger.info('transforming tokens to lowercase')
@@ -656,6 +953,11 @@ class TMPreproc:
         return self
 
     def stem(self):
+        """
+        Apply stemming to all tokens using function :attr:`~TMPreproc.stemmer`.
+
+        :return: this instance
+        """
         self._require_no_ngrams_as_tokens()
 
         self._invalidate_workers_tokens()
@@ -667,12 +969,16 @@ class TMPreproc:
 
     def pos_tag(self):
         """
-        Apply Part-of-Speech (POS) tagging to all documents. POS tags can then be retrieved via `tokens_with_metadata`
-        or `tokens_with_pos_tags` properties or the `get_tokens()` method.
+        Apply Part-of-Speech (POS) tagging to all tokens using :attr:`~TMPreproc.pos_tagger`. POS tags can then be
+        retrieved via :attr:`~TMPreproc.tokens_with_metadata` or :attr:`~TMPreproc.tokens_with_pos_tags` properties or
+        the :meth:`~TMPreproc.get_tokens` method.
 
-        POS tagging so far only works for English and German. The English tagger uses the Penn Treebank tagset
-        (https://ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html), the
-        German tagger uses STTS (http://www.ims.uni-stuttgart.de/forschung/ressourcen/lexika/TagSets/stts-table.html).
+        With the default POS tagging function, tagging so far only works for English and German. The English tagger
+        uses the `Penn Treebank tagset <https://ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html>`_,
+        the German tagger uses
+        `STTS <http://www.ims.uni-stuttgart.de/forschung/ressourcen/lexika/TagSets/stts-table.html>`_.
+
+        :return: this instance
         """
         self._require_no_ngrams_as_tokens()
 
@@ -683,6 +989,11 @@ class TMPreproc:
         return self
 
     def lemmatize(self):
+        """
+        Lemmatize tokens using function :attr:`~TMPreproc.lemmatizer`.
+
+        :return: this instance
+        """
         self._require_pos_tags()
         self._require_no_ngrams_as_tokens()
 
@@ -694,6 +1005,18 @@ class TMPreproc:
         return self
 
     def expand_compound_tokens(self, split_chars=('-',), split_on_len=2, split_on_casechange=False):
+        """
+        Expand compound tokens like "US-Student" to "US" and "Student". Use `split_chars` to determine possible
+        split points and/or case changes (e.g. "USstudent") when setting `split_on_casechange` to True.
+        The minimum length of the split sub-strings must be `split_on_len`.
+
+        .. seealso:: :func:`~tmtoolkit.preprocess.expand_compound_token`
+
+        :param split_chars: possibly split on these characters
+        :param split_on_len: ensure that split sub-strings have at least this length
+        :param split_on_casechange: also split on case changes
+        :return: this instance
+        """
         self._require_no_ngrams_as_tokens()
 
         self._invalidate_workers_tokens()
@@ -707,9 +1030,23 @@ class TMPreproc:
         return self
 
     def remove_special_chars_in_tokens(self):
+        """
+        Remove everything that is deemed a "special character", i.e. everything in :attr:`~TMPreproc.special_chars`
+        from all tokens. Be default, this will remove all characters listed in :func:`strings.punctuation`.
+
+        :return: this instance
+        """
         return self.remove_chars_in_tokens(self.special_chars)
 
     def remove_chars_in_tokens(self, chars):
+        """
+        Remove all characters listed in `chars` from all tokens.
+
+        :param chars: list of characters to remove
+        :return: this instance
+        """
+        require_listlike(chars)
+
         self._invalidate_workers_tokens()
 
         logger.info('removing characters in tokens')
@@ -719,6 +1056,19 @@ class TMPreproc:
 
     def clean_tokens(self, remove_punct=True, remove_stopwords=True, remove_empty=True,
                      remove_shorter_than=None, remove_longer_than=None, remove_numbers=False):
+        """
+        Clean tokens by removing a certain, configurable subset of them.
+
+        :param remove_punct: remove all tokens that intersect with punctuation tokens from
+                             :attr:`~TMPreproc.punctuation`
+        :param remove_stopwords: remove all tokens that intersect with stopword tokens from :attr:`~TMPreproc.stopwords`
+        :param remove_empty: remove all empty string ``""`` tokens
+        :param remove_shorter_than: remove all tokens shorter than this length
+        :param remove_longer_than: remove all tokens longer than this length
+        :param remove_numbers: remove all tokens that are "numeric" according to the NumPy function
+                               :func:`numpy.char.isnumeric`
+        :return: this instance
+        """
         tokens_to_remove = [''] if remove_empty else []
 
         if remove_punct:
@@ -738,7 +1088,66 @@ class TMPreproc:
 
         return self
 
-    def filter_tokens(self, search_tokens, match_type='exact', ignore_case=False, glob_method='match', inverse=False):
+    def filter_tokens_by_mask(self, mask, inverse=False):
+        """
+        Filter tokens according to a binary mask specified by `mask`.
+
+        .. seealso:: :meth:`~TMPreproc.remove_tokens_by_mask`
+
+        :param mask: a dict containing a mask list for each document; each mask list contains boolean values for
+                     each token in that document, where `True` means keeping that token and `False` means removing it
+        :param inverse: inverse the mask for filtering, i.e. keep all tokens with a mask set to `False` and remove all
+                        those with `True`
+        :return: this instance
+        """
+
+        self._invalidate_workers_tokens()
+
+        logger.info('filtering tokens by mask')
+        self._send_task_to_workers('filter_tokens_by_mask', mask=mask, inverse=inverse)
+
+        return self
+
+    def remove_tokens_by_mask(self, mask):
+        """
+        Remove tokens according to a binary mask specified by `mask`.
+
+        .. seealso:: :meth:`~TMPreproc.filter_tokens_by_mask`
+
+        :param mask: a dict containing a mask list for each document; each mask list contains boolean values for
+                     each token in that document, where `False` means keeping that token and `True` means removing it
+        :return: this instance
+        """
+
+        return self.filter_tokens_by_mask(mask, inverse=True)
+
+    def filter_tokens(self, search_tokens, by_meta=None, match_type='exact', ignore_case=False, glob_method='match',
+                      inverse=False):
+        """
+        Filter tokens according to search pattern(s) `search_tokens` and several matching options. Only those tokens
+        are retained that match the search criteria unless you set ``inverse=True``, which will *remove* all tokens
+        that match the search criteria (which is the same as calling :meth:`~TMPreproc.remove_tokens`).
+
+        .. seealso:: :meth:`~TMPreproc.remove_tokens` and :func:`~tmtoolkit.preprocess.token_match`
+
+        :param search_tokens: single string or list of strings that specify the search pattern(s)
+        :param by_meta: if not None, this should be a string of a meta data key; this meta data will then be
+                        used for matching instead of the tokens in `docs`
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :param inverse: inverse the match results for filtering (i.e. *remove* all tokens that match the search
+                        criteria)
+        :return: this instance
+        """
+        if by_meta and by_meta not in self.get_available_metadata_keys():
+            raise ValueError('meta data key `%s` is not available' % by_meta)
+
         self._check_filter_args(match_type=match_type, glob_method=glob_method)
 
         self._invalidate_workers_tokens()
@@ -749,17 +1158,60 @@ class TMPreproc:
                                    match_type=match_type,
                                    ignore_case=ignore_case,
                                    glob_method=glob_method,
-                                   inverse=inverse)
+                                   inverse=inverse,
+                                   by_meta=by_meta)
 
         return self
 
-    def remove_tokens(self, search_tokens, match_type='exact', ignore_case=False, glob_method='match'):
+    def remove_tokens(self, search_tokens, by_meta=None, match_type='exact', ignore_case=False, glob_method='match'):
+        """
+        This is a shortcut for the :meth:`~TMPreproc.filter_tokens` method with ``inverse=True``, i.e. *remove* all
+        tokens that match the search criteria).
+
+        .. seealso:: :meth:`~TMPreproc.filter_tokens` and :func:`~tmtoolkit.preprocess.token_match`
+
+        :param search_tokens: single string or list of strings that specify the search pattern(s)
+        :param by_meta: if not None, this should be a string of a meta data key; this meta data will then be
+                        used for matching instead of the tokens in `docs`
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :return: this instance
+        """
         return self.filter_tokens(search_tokens=search_tokens, match_type=match_type,
                                   ignore_case=ignore_case, glob_method=glob_method,
-                                  inverse=True)
+                                  by_meta=by_meta, inverse=True)
 
-    def filter_documents(self, search_tokens, matches_threshold=1, match_type='exact', ignore_case=False,
-                         glob_method='match', inverse=False):
+    def filter_documents(self, search_tokens, by_meta=None, matches_threshold=1, match_type='exact', ignore_case=False,
+                         glob_method='match', inverse_result=False, inverse_matches=False):
+        """
+        This method is similar to `filter_tokens` but applies at document level. For each document, the number of
+        matches is counted. If it is at least `matches_threshold` the document is retained, otherwise removed. If
+        `inverse_result` is True, then documents that meet the threshold are *removed*.
+
+        .. seealso:: :meth:`~TMPreproc.remove_documents`
+
+        :param search_tokens: single string or list of strings that specify the search pattern(s)
+        :param by_meta: if not None, this should be a string of a meta data key; this meta data will then be
+                        used for matching instead of the tokens in `docs`
+        :param matches_threshold: the minimum number of matches required per document
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :param inverse_result: inverse the threshold comparison result
+        :param inverse_matches: inverse the match results for filtering
+        :return: this instance
+        """
         self._check_filter_args(match_type=match_type, glob_method=glob_method)
 
         n_docs_orig = self.n_docs
@@ -770,35 +1222,61 @@ class TMPreproc:
         logger.info('filtering %d documents' % n_docs_orig)
         self._send_task_to_workers('filter_documents',
                                    search_tokens=search_tokens,
+                                   by_meta=by_meta,
                                    matches_threshold=matches_threshold,
                                    match_type=match_type,
                                    ignore_case=ignore_case,
                                    glob_method=glob_method,
-                                   inverse=inverse)
+                                   inverse_result=inverse_result,
+                                   inverse_matches=inverse_matches)
 
         return self
 
-    def remove_documents(self, search_token, matches_threshold=1, match_type='exact', ignore_case=False,
-                         glob_method='match'):
-        return self.filter_documents(search_tokens=search_token, matches_threshold=matches_threshold,
+    def remove_documents(self, search_tokens, by_meta=None, matches_threshold=1, match_type='exact', ignore_case=False,
+                         glob_method='match', inverse_matches=False):
+        """
+        This is a shortcut for the `filter_documents` method with ``inverse_result=True``, i.e. *remove* all
+        documents that meet the token matching threshold.
+
+        .. seealso:: :meth:`~TMPreproc.filter_documents`
+
+        :param search_tokens: single string or list of strings that specify the search pattern(s)
+        :param by_meta: if not None, this should be a string of a meta data key; this meta data will then be
+                        used for matching instead of the tokens in `docs`
+        :param matches_threshold: the minimum number of matches required per document
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :param inverse_matches: inverse the match results for filtering
+        :return: this instance
+        """
+        return self.filter_documents(search_tokens=search_tokens, by_meta=by_meta, matches_threshold=matches_threshold,
                                      match_type=match_type, ignore_case=ignore_case, glob_method=glob_method,
-                                     inverse=True)
+                                     inverse_matches=inverse_matches, inverse_result=True)
 
     def filter_documents_by_name(self, name_patterns, match_type='exact', ignore_case=False, glob_method='match',
                                  inverse=False):
         """
         Filter documents by their name (i.e. document label). Keep all documents whose name matches `name_pattern`
         according to additional matching options. If `inverse` is True, drop all those documents whose name matches,
-        which is the same as calling `remove_documents_by_name`.
+        which is the same as calling :meth:`~TMPreproc.remove_documents_by_name`.
+
         :param name_patterns: either single search string or sequence of search strings
-        :param match_type: One of: 'exact', 'regex', 'glob'. If 'regex', `search_token` must be RE pattern. If `glob`,
-                           `search_token` must be a "glob" pattern like "hello w*"
-                           (see https://github.com/metagriffin/globre).
-        :param ignore_case: If True, ignore case for matching.
-        :param glob_method: If `match_type` is 'glob', use this glob method. Must be 'match' or 'search' (similar
-                            behavior as Python's `re.match` or `re.search`).
-        :param inverse: Invert the matching results.
-        :return: this TMPreproc instance
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :param inverse: invert the matching results
+        :return: this instance
         """
         self._check_filter_args(match_type=match_type, glob_method=glob_method)
 
@@ -819,15 +1297,19 @@ class TMPreproc:
 
     def remove_documents_by_name(self, name_patterns, match_type='exact', ignore_case=False, glob_method='match'):
         """
-        Same as `filter_documents_by_name` with `inverse=True`: drop all those documents whose name matches.
+        Same as :meth:`~TMPreproc.filter_documents_by_name` with ``inverse=True``: drop all those documents whose
+        name match.
+
         :param name_patterns: either single search string or sequence of search strings
-        :param match_type: One of: 'exact', 'regex', 'glob'. If 'regex', `search_token` must be RE pattern. If `glob`,
-                           `search_token` must be a "glob" pattern like "hello w*"
-                           (see https://github.com/metagriffin/globre).
-        :param ignore_case: If True, ignore case for matching.
-        :param glob_method: If `match_type` is 'glob', use this glob method. Must be 'match' or 'search' (similar
-                            behavior as Python's `re.match` or `re.search`).
-        :return: this TMPreproc instance
+        :param match_type: the type of matching that is performed: ``'exact'`` does exact string matching (optionally
+                           ignoring character case if ``ignore_case=True`` is set); ``'regex'`` treats ``search_tokens``
+                           as regular expressions to match the tokens against; ``'glob'`` uses "glob patterns" like
+                           ``"politic*"`` which matches for example "politic", "politics" or ""politician" (see
+                           `globre package <https://pypi.org/project/globre/>`_)
+        :param ignore_case: ignore character case (applies to all three match types)
+        :param glob_method: if `match_type` is ``'glob'``, use either ``'search'`` or ``'match'`` as glob method
+                            (has similar implications as Python's ``re.search`` vs. ``re.match``)
+        :return: this instance
         """
         return self.filter_documents_by_name(name_patterns, match_type=match_type,
                                              ignore_case=ignore_case, glob_method=glob_method,
@@ -838,11 +1320,17 @@ class TMPreproc:
         Filter tokens for a specific POS tag (if `required_pos` is a string) or several POS tags (if `required_pos`
         is a list/tuple/set of strings). The POS tag depends on the tagset used during tagging. If `simplify_pos` is
         True, then the tags are matched to the following simplified forms:
-        - 'N' for nouns
-        - 'V' for verbs
-        - 'ADJ' for adjectives
-        - 'ADV' for adverbs
-        - None for all other
+
+        * ``'N'`` for nouns
+        * ``'V'`` for verbs
+        * ``'ADJ'`` for adjectives
+        * ``'ADV'`` for adverbs
+        * ``None`` for all other
+
+        :param required_pos: single string or list of strings with POS tag(s) used for filtering
+        :param simplify_pos: before matching simplify POS tags in documents to forms shown above
+        :param inverse: inverse the matching results, i.e. *remove* tokens that match the POS tag
+        :return: this instance
         """
         if not isinstance(required_pos, (tuple, list, set, str)) \
                 and required_pos is not None:
@@ -862,36 +1350,62 @@ class TMPreproc:
         return self
 
     def remove_tokens_by_doc_frequency(self, which, df_threshold, absolute=False):
-        blacklist = remove_tokens_by_doc_frequency([doc['token'] for doc in self._workers_tokens.values()], which,
-                                                   df_threshold=df_threshold, absolute=absolute, return_blacklist=True)
+        """
+        Remove tokens according to their document frequency.
 
-        if blacklist:
-            self._invalidate_workers_tokens()
+        :param which: which threshold comparison to use: either ``'common'``, ``'>'``, ``'>='`` which means that tokens
+                      with higher document freq. than (or equal to) `df_threshold` will be removed;
+                      or ``'uncommon'``, ``'<'``, ``'<='`` which means that tokens with lower document freq. than
+                      (or equal to) `df_threshold` will be removed
+        :param df_threshold: document frequency threshold value
+        :param absolute: if True, use absolute document frequency (i.e. number of times token X occurs at least once
+                         in a document), otherwise use relative document frequency (normalized by number of documents)
+        :return: this instance
+        """
+        mask = remove_tokens_by_doc_frequency([doc['token'] for doc in self._workers_tokens.values()], which,
+                                               df_threshold=df_threshold, absolute=absolute, return_mask=True)
 
-            logger.debug('will remove the following %d tokens: %s' % (len(blacklist), blacklist))
-            self._send_task_to_workers('filter_tokens',
-                                       search_tokens=blacklist,
-                                       match_type='exact',
-                                       ignore_case=False,
-                                       glob_method='match',
-                                       inverse=True)
+        if sum(sum(dmsk) for dmsk in mask) > 0:
+            self.remove_tokens_by_mask(dict(zip(self._workers_tokens.keys(), mask)))
 
         return self
 
     def remove_common_tokens(self, df_threshold, absolute=False):
+        """
+        Remove tokens with document frequency greater than or equal to `df_threshold`.
+
+        :param df_threshold: document frequency threshold value
+        :param absolute: if True, use absolute document frequency (i.e. number of times token X occurs at least once
+                         in a document), otherwise use relative document frequency (normalized by number of documents)
+        :return: this instance
+        """
         return self.remove_tokens_by_doc_frequency('common', df_threshold=df_threshold, absolute=absolute)
 
     def remove_uncommon_tokens(self, df_threshold, absolute=False):
+        """
+        Remove tokens with document frequency lesser than or equal to `df_threshold`.
+
+        :param df_threshold: document frequency threshold value
+        :param absolute: if True, use absolute document frequency (i.e. number of times token X occurs at least once
+                         in a document), otherwise use relative document frequency (normalized by number of documents)
+        :return: this instance
+        """
+
         return self.remove_tokens_by_doc_frequency('uncommon', df_threshold=df_threshold, absolute=absolute)
 
-    def apply_custom_filter(self, filter_func, to_tokens_dataframe=False):
+    def apply_custom_filter(self, filter_func, to_tokens_datatable=False):
         """
         Apply a custom filter function `filter_func` to all tokens or tokens dataframe.
-        `filter_func` must accept a single parameter: a dictionary of structure `{<doc_label>: <tokens list>}` as from
-        `.tokens` if `to_tokens_dataframe` is False or a data frame as from `tokens_datatable`. It must return a result
-        with the same structure.
+        `filter_func` must accept a single parameter: a dictionary of structure ``{<doc_label>: <tokens list>}`` as from
+        :attr:`~TMPreproc.tokens` if `to_tokens_dataframe` is False or a datatable Frame as from
+        :attr:`~TMPreproc.tokens_datatable`. It must return a result with the same structure.
 
-        This function can only be run on a single process, hence it could be slow for large corpora.
+        :param filter_func: filter function to apply to all tokens or tokens dataframe
+        :param to_tokens_datatable: if True, pass datatable as from :attr:`~TMPreproc.tokens_datatable` to
+                                    `filter_func`, otherwise pass dict as from :attr:`~TMPreproc.tokens` to
+                                    `filter_func`
+
+        .. warning:: This function can only be run on a single process, hence it could be slow for large corpora.
         """
 
         # Because it is not possible to send a function to the workers, all tokens must be fetched from the workers
@@ -903,8 +1417,8 @@ class TMPreproc:
 
         self._invalidate_workers_tokens()
 
-        if to_tokens_dataframe:
-            logger.info('applying custom filter function to tokens data frame')
+        if to_tokens_datatable:
+            logger.info('applying custom filter function to tokens datatable')
             data = self.tokens_datatable
         else:
             logger.info('applying custom filter function to tokens')
@@ -912,26 +1426,27 @@ class TMPreproc:
 
         res = filter_func(data)
 
-        if to_tokens_dataframe:
+        if to_tokens_datatable:
             self.load_tokens_datatable(res)
         else:
             self.load_tokens(res)
 
         return self
 
-    def get_dtm(self, as_data_table=False, as_data_frame=False, dtype=None):
+    def get_dtm(self, as_datatable=False, as_data_frame=False, dtype=None):
         """
         Generate a sparse document-term-matrix (DTM) for the current tokens with rows representing documents according
-        to `self.doc_labels` and columns representing tokens according to `self.vocabulary`.
-        :param as_data_table: Return result as data table with document labels in '_doc' column and vocabulary as
+        to :attr:`~TMPreproc.doc_labels` and columns representing tokens according to :attr:`~TMPreproc.vocabulary`.
+
+        :param as_datatable: Return result as data table with document labels in '_doc' column and vocabulary as
                               column names
         :param as_data_frame: Return result as pandas data frame with document labels in index and vocabulary as
                               column names
-        :param dtype: optionally specify a DTM data type. by default it is 32bit integer
-        :return: either a sparse document-term-matrix (in CSR format) or a datatable, depending on `as_data_table`
+        :param dtype: optionally specify a DTM data type; by default it is 32bit integer
+        :return: either a sparse document-term-matrix (in CSR format) or a datatable or a pandas DataFrame
         """
-        if as_data_table and as_data_frame:
-            raise ValueError('you cannot set both `as_data_table` and `as_data_frame` to True')
+        if as_datatable and as_data_frame:
+            raise ValueError('you cannot set both `as_datatable` and `as_data_frame` to True')
 
         if self._cur_dtm is None:
             logger.info('generating DTM')
@@ -955,7 +1470,7 @@ class TMPreproc:
         else:
             vocab = self.get_vocabulary()
 
-        if as_data_table:
+        if as_datatable:
             return dtm_to_datatable(self._cur_dtm.todense(), self.doc_labels, vocab)
         elif as_data_frame:
             return dtm_to_dataframe(self._cur_dtm.todense(), self.doc_labels, vocab)
@@ -964,6 +1479,10 @@ class TMPreproc:
 
     @property
     def _workers_tokens(self):
+        """
+        Fetch tokens from workers and return as dict with document label -> document tokens mapping.
+        Returns cached `_cur_workers_tokens` if it is not None.
+        """
         if self._cur_workers_tokens is not None:
             return self._cur_workers_tokens
 
@@ -976,11 +1495,17 @@ class TMPreproc:
 
     @property
     def _workers_vocab_doc_frequencies(self):
+        """
+        Fetch document frequency for each vocabulary token from workers and return as :class:`collections.Counter()`
+        instance with token -> doc. frequency mapping.
+        Returns cached `_cur_workers_vocab_doc_freqs` if it is not None.
+        """
         if self._cur_workers_vocab_doc_freqs is not None:
             return self._cur_workers_vocab_doc_freqs
 
         workers_doc_freqs = self._get_results_seq_from_workers('get_vocab_doc_frequencies')
 
+        # sum up the individual workers' results
         self._cur_workers_vocab_doc_freqs = Counter()
         for doc_freqs in workers_doc_freqs:
             self._cur_workers_vocab_doc_freqs.update(doc_freqs)
@@ -989,6 +1514,10 @@ class TMPreproc:
 
     @property
     def _workers_ngrams(self):
+        """
+        Fetch ngrams from workers and return as dict with document label -> document ngrams mapping.
+        Returns cached `_cur_workers_ngrams` if it is not None.
+        """
         if self._cur_workers_ngrams is not None:
             return self._cur_workers_ngrams
 
@@ -1001,6 +1530,7 @@ class TMPreproc:
         return self._cur_workers_ngrams
 
     def _add_metadata(self, task, key, data, default):
+        """Add metadata either per document or per token."""
         if not isinstance(data, dict):
             raise ValueError('`data` must be a dict')
 
@@ -1040,6 +1570,10 @@ class TMPreproc:
             self._send_task_to_workers(task, key=key, data=data, default=default)
 
     def _create_state_object(self, deepcopy_attrs):
+        """
+        Create a state object as dict for the current instance by (deep) copying this instance' attributes and fetching
+        each workers' state.
+        """
         state_attrs = {}
         attr_blacklist = ('tasks_queues', 'results_queue',
                           'workers', 'n_workers', 'lemmatizer')
@@ -1143,15 +1677,22 @@ class TMPreproc:
         self.n_workers = len(self.workers)
 
     def _send_task_to_workers(self, task, **kwargs):
+        """
+        Send the task identified by string `task` to all workers passing additional task parameters via `kwargs`.
+        Run these tasks and parallel and wait until all are finished.
+        """
         if not (self.tasks_queues and self.workers):
             return
 
-        shutdown = task is None
-        task_item = None if shutdown else (task, kwargs)
+        shutdown = task is None                            # "None" is used to signal worker shutdown
+        task_item = None if shutdown else (task, kwargs)   # a task item consists of the task name and its parameters
 
         logger.debug('sending task `%s` to all workers' % task)
 
+        # put the task in each worker's task queue
         [q.put(task_item) for q in self.tasks_queues]
+
+        # run join() on each worker's task queue in order to wait for the tasks to finish
         [q.join() for q in self.tasks_queues]
 
         if shutdown:
@@ -1163,26 +1704,21 @@ class TMPreproc:
             self.n_workers = 0
 
     def _get_results_seq_from_workers(self, task, **kwargs):
+        """
+        Send the task identified by string `task` to all workers passing additional task parameters via `kwargs`. After
+        the tasks are finished, collect the results in a list with individual results for each worker and return it.
+        """
         logger.debug('getting results sequence for task `%s` from all workers' % task)
         self._send_task_to_workers(task, **kwargs)
 
         return [self.results_queue.get() for _ in range(self.n_workers)]
 
-    def _get_results_per_worker(self, task, **kwargs):
-        logger.debug('getting results per worker for task `%s` from all workers' % task)
-        self._send_task_to_workers(task, **kwargs)
-
-        res = [self.results_queue.get() for _ in range(self.n_workers)]
-
-        if res and set(map(len, res)) != {2}:
-            raise RuntimeError('all workers must return a 2-tuple as result')
-
-        return dict(res)
-
     def _invalidate_docs_info(self):
+        """Invalidate cached data related to document information (such as document labels)."""
         self._cur_doc_labels = None
 
     def _invalidate_workers_tokens(self):
+        """Invalidate cached data related to worker tokens."""
         self._cur_workers_tokens = None
         self._cur_workers_vocab = None
         self._cur_workers_vocab_doc_freqs = None
@@ -1190,6 +1726,7 @@ class TMPreproc:
         self._cur_dtm = None
 
     def _invalidate_workers_ngrams(self):
+        """Invalidate cached data related to worker ngrams."""
         self._cur_workers_ngrams = None
         self._cur_workers_vocab = None
         self._cur_workers_vocab_doc_freqs = None
