@@ -1,9 +1,8 @@
 """
 Common constants and functions
-
-Markus Konrad <markus.konrad@wzb.eu>
 """
 
 
 DEFAULT_TOPIC_NAME_FMT = 'topic_{i1}'
 DEFAULT_RANK_NAME_FMT = 'rank_{i1}'
+DEFAULT_VALUE_FORMAT = '{lbl} ({val:.4})'
